+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.0] - 15/11/2024
+### Add rect and drawing function utility
+- Add functions to draw cross and border on a rect
+- Add rect utility class to handle rect manipulation
+
 ## [0.1.0] - 15/11/2024
 ### First alpha release
 - Add singleton class to extend